refactor: 페어 리뷰 반영 Store 전반 리팩터링
